requirements.txt
pillow==10.2.0
pytesseract==0.3.10